move pdf helper alongside `Page`
