+++ src/utils.rs
@@ -713,17 +713,3 @@ pub fn fit_bounds(width: f32, height: f32, src: Rect, dst: Rect) -> (Rect, Rect)
 
   (src, dst)
 }
-
-//
-// PDF creation
-//
-
-use skia_safe::{pdf, Document};
-
-pub fn pdf_document(quality:f32, density:f32) -> Document{
-  let mut meta = pdf::Metadata::default();
-  meta.producer = "Skia Canvas <https://github.com/samizdatco/skia-canvas>".to_string();
-  meta.encoding_quality = Some((quality*100.0) as i32);
-  meta.raster_dpi = Some(density * 72.0);
-  pdf::new_document(Some(&meta))
-}
